+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.4.4]
+
+* Fix wrong caret hegiht and postion
+* Make Ios/Android caret the same height 
+
 ## [0.4.3]
 
 * Fix kMinInteractiveSize is missing in high version of flutter